Decouple if-arrow and if-then expression
